# api/docs/swagger.yaml
basePath: /
definitions:
...
  controllers.ProductInput:
    properties:
      category:
        type: string
      description:
        type: string
      image_url:
        type: string
      name:
        type: string
      purchase_price:
        type: number
      selling_price:
        type: number
      stock:
        type: integer
    required:
    - name
    - purchase_price
    - selling_price
    - stock
    type: object
...
  controllers.TransactionExpenseInput:
    properties:
      amount:
        type: number
    type: object
  controllers.TransactionSaleInput:
    properties:
      product_id:
        type: integer
      quantity:
        type: integer
    type: object
  controllers.TransactionWithdrawalInput:
    properties:
      amount:
        type: number
    type: object
...
  services.TransactionResponse:
    properties:
      amount:
        type: number
      created_at:
        type: string
      id:
        type: integer
      product_id:
        type: integer
      quantity:
        type: integer
      type:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: Backend multi-tenant for electronics shops
  title: MultiShop API
  version: "1.0"
paths:
...
  /api/products:
    get:
      description: Returns all products belonging to the authenticated user's shop.
        SuperAdmin can see PurchasePrice, Admin cannot.
      produces:
      - application/json
      responses:
        "200":
          description: List of products
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get all products of the current shop
      tags:
      - Products
    post:
      consumes:
      - application/json
      description: Allows an Admin or SuperAdmin to create a new product in their
        shop. Product name must be unique per shop.
      parameters:
      - description: Product creation payload
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/controllers.ProductInput'
      produces:
      - application/json
      responses:
        "201":
          description: Product created successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid input
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create a new product
      tags:
      - Products
  /api/products/{id}:
    delete:
      description: Allows a SuperAdmin to delete a product belonging to their shop.
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Product deleted successfully
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden - SuperAdmin only
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Delete a product
      tags:
      - Products
    put:
      consumes:
      - application/json
      description: Allows Admin or SuperAdmin to update a product belonging to their
        shop. Product name must remain unique. Only SuperAdmin can update PurchasePrice.
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated product payload
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/controllers.ProductInput'
      produces:
      - application/json
      responses:
        "200":
          description: Product updated successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid input or product not found
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update an existing product
      tags:
      - Products
...
  /api/transactions:
    get:
      description: Returns all transactions (Sale, Expense, Withdrawal) belonging
        to the authenticated user's shop.
      produces:
      - application/json
      responses:
        "200":
          description: List of transactions
          schema:
            items:
              $ref: '#/definitions/services.TransactionResponse'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get all transactions of the current shop
      tags:
      - Transactions
  /api/transactions/{id}:
    delete:
      description: Allows Admin or SuperAdmin to delete a transaction belonging to
        their shop. If the transaction is a Sale, the product stock is restored automatically.
      parameters:
      - description: Transaction ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Transaction deleted successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Transaction not found
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Delete a transaction
      tags:
      - Transactions
  /api/transactions/expense:
    post:
      consumes:
      - application/json
      description: Allows Admin or SuperAdmin to create an Expense transaction. The
        request must include the amount of the expense.
      parameters:
      - description: Transaction payload
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/controllers.TransactionExpenseInput'
      produces:
      - application/json
      responses:
        "201":
          description: Transaction created successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid input
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create a new Expense transaction
      tags:
      - Transactions
  /api/transactions/sale:
    post:
      consumes:
      - application/json
      description: Allows Admin or SuperAdmin to create a Sale transaction. The request
        must include the product ID and quantity. The system will automatically calculate
        the total amount based on the product price and quantity, and reduce the product
        stock accordingly.
      parameters:
      - description: Transaction payload
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/controllers.TransactionSaleInput'
      produces:
      - application/json
      responses:
        "201":
          description: Transaction created successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid input, insufficient stock, or product not found
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create a new Sale transaction
      tags:
      - Transactions
  /api/transactions/withdrawal:
    post:
      consumes:
      - application/json
      description: Allows Admin or SuperAdmin to create a Withdrawal transaction.
        The request must include the amount of the withdrawal.
      parameters:
      - description: Transaction payload
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/controllers.TransactionWithdrawalInput'
      produces:
      - application/json
      responses:
        "201":
          description: Transaction created successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid input
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create a new Withdrawal transaction
      tags:
      - Transactions
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"